Turret focusing
Considering restarting the development to make t easier to follwo but that should come later. The blueprints are vvery messy at this point and require substantial changes. Discovered another way to look at altering and moving the turret but that requires using the pawn class so i may need to transfer some of bens base class to a pawn base class.