Further improvements to enemy grappling
Need to improve the grapple logic when launching the character up and calculating the land location to be in front of the player.
Grappling is now based on the enemy size, only small enemies are grappled to the player.
Grapple cable currently does not update with enemy movement.

Note: If grapples don't appear properly, its because of the whip volume, as always :)